dump: Rework command to encode
In order to give the command more utility and play well with pipeline
idea of the vegeta toolchain we decided per discussion[0] to change the
semantics of dump. We offer now a general encode command which can
transofrm any input encoding to any of the supported output encodings.
We anticipate that in conjunction with advocated auxiliary tools, like\
jq, a wide range of use-cases can be covered without extending vegeta
itself.

[0] https://github.com/tsenart/vegeta/issues/302#issuecomment-403935450<Paste>

Closes #302